Added new graph library and show a dummy graph
Have shown a graph with dummy data. Will begin to implement backend php now to retrieve real data as well as customise the Graph component so it will take inputs to show the different graph types. Will also improve physical look but this is easy and I would like the teams input.